[FL-1595] Add EMV tags (#625)
* nfc: add expiration date tag to emv parser
* nfc: add expiration date save and display
* nfc: add long apdu test command

+++ applications/nfc/nfc_device.h
@@ -42,7 +42,7 @@ typedef struct {
     uint16_t aid_len;
     uint8_t number[8];
     uint8_t exp_mon;
-    uint16_t exp_year;
+    uint8_t exp_year;
     char cardholder[32];
 } NfcEmvData;
